:data update:
With new data comes new responsibilities.
In the dataset creation, one can find out how the dataset was curated and
some example plots of it.
You can reuse the jupyter notebook to make your own dataset (given you
have the same initial pulse dicitonaries)

Following changes to the dataset, we have to change the way the data is
loaded and read with the model, i.e., lightning module, DIVA, etc.,
needed to be updated to reflect these changes.

Originally, we only had the masks, profiles, controls, and raddi. Now
there is much more included! See dataset creation for a better outline
of what is there. Also a pdf will be at a later date to showcase the
dataset.