:physics update:
	Here we add a few plotting routines in Pulse Comparisons.
	This will serve as a baseline for devising physics experiments,
i.e., when we want to compare two pulses that are similar in regime, but
vary in one dimensions.
	Then we add to the model a bit of physics! This is that the
stored energy should be equal between pulses. Stored energy simply
calculated by static pressure sum.
	This is coupled with passing the normalization constants for
profiels to the loss function, as we need to calculate static stored energy based on the real profiles (and multiply by boltzmann constant), hence the change in the experiment file.
	Visualizations are for the dataset.